mejora de logica

+++ src/main/java/com/gamm/hotel_api/dto/DetalleReservaDTO.java
@@ -2,6 +2,8 @@ package com.gamm.hotel_api.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +14,7 @@ public class DetalleReservaDTO {
   private Integer idServicio;
   private String nombreServicio;
   private Integer cantidad;
-  private Double precio;
-  private Double subtotal;
+  private BigDecimal precio;
+  private BigDecimal subtotal;
 }
